refactor(settings): M-1b — extract shared components from SettingsView.swift (UI-002)
Audit M-1b: continue decomposing SettingsView.swift. Moves shared scaffolds and
form components to their own files under Settings/v2/Components/.

Extracted:
- Components/SettingsScaffolds.swift (94 lines): SettingsDetailScaffold,
  SettingsSectionCard, SettingsValueRow, SettingsSupportingText
- Components/SettingsFormComponents.swift (165 lines): SettingsActionTrailing,
  SettingsActionLabel, SettingsSelectionTile, SettingsChoiceGrid,
  SettingsNumericFieldRow, SettingsSliderRow

SettingsView.swift drops 649 → 401 lines (~38% further reduction).
Total since M-1a start: 1170 → 401 lines (~66%).

Visibility bumped on SettingsHomeHeader from private → internal (referenced
by SettingsDetailScaffold in the new Components file). All other types
already internal from M-1a.

No behaviour changes — pure structural decomposition. Build green.

Plan: docs/superpowers/plans/2026-04-19-m1-settings-v3-decomposition.md
